# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Wattson Backend API
  version: 0.1.0
  description: The small contract between the Wattson web app and planning service.
paths:
...
  /v1/demo/sites/{site_id}/scenario:
    get:
      operationId: getDemoScenario
      parameters:
        - name: site_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The site's validated 24-hour planning scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: The site does not have a seeded scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sites:
    get:
      operationId: listSites
      description: Lists the latest persisted scenario for each site with capacity and latest plan metrics.
      responses:
        "200":
          description: The persisted sites.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SiteSummary" }
    post:
      operationId: createSite
      description: Creates a site with a starter solar grid and a validated 24-hour scenario.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteRequest" }
      responses:
        "201":
          description: The created site summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteSummary" }
        "400":
          description: The site input is invalid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: A site with this generated ID already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sites/{site_id}:
    get:
      operationId: getSite
      parameters:
        - name: site_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The site, its current scenario, counts, and latest plan summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteOverview" }
        "404":
          description: The site does not exist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sites/{site_id}/forecast:
    get:
      operationId: getSiteForecast
      description: Returns a live or cached seven-day Open-Meteo forecast and asset-specific operating actions.
      parameters:
        - name: site_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The normalized weather forecast. A stale cache is returned only when Open-Meteo fails.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeatherForecast" }
        "404":
          description: The site does not exist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Open-Meteo failed and no cached forecast exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/scenarios/{scenario_id}:
    get:
      operationId: getScenario
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The current revision of the scenario input for planning.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: replaceScenario
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Scenario"
      responses:
        "200":
          description: The full validated replacement scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The body is invalid or its ID does not match the path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteScenario
      description: Deletes the scenario and all plan runs that belong to it.
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: The scenario and its plan runs were deleted.
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/connections:
    put:
      operationId: replaceScenarioConnections
      description: Replaces the complete directed connection graph and increments the scenario revision.
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionsRequest"
      responses:
        "200":
          description: The updated full scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The connection graph is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios:
    get:
      operationId: listScenarios
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: The available scenario revisions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScenarioSummary"
        "400":
          description: A pagination value is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createScenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Scenario"
      responses:
        "201":
          description: The validated scenario.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The scenario body is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A scenario uses this ID already.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/signals/{signal_id}:
    put:
      operationId: replaceScenarioSignalValues
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
        - name: signal_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalValuesUpdate"
      responses:
        "200":
          description: The scenario with the validated signal values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The values make the scenario invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario or signal does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/initial-state:
    put:
      operationId: replaceScenarioInitialState
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitialState"
      responses:
        "200":
          description: The scenario with the validated battery and fuel state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The initial state makes the scenario invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/events:
    post:
      operationId: createScenarioEvent
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioEvent"
      responses:
        "201":
          description: The scenario with the new event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "400":
          description: The event makes the scenario invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: An event uses this ID already.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/events/{event_id}:
    delete:
      operationId: deleteScenarioEvent
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
        - name: event_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: The event was removed from the scenario.
        "404":
          description: The scenario or event does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/plan-runs:
    post:
      operationId: createPlanRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanningRequest"
      responses:
        "201":
          description: The completed or infeasible plan run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanRun"
        "400":
          description: The request body or parent run is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The scenario does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The parent baseline uses a different scenario revision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: The scenario or active event selection cannot be planned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/plan-runs/{run_id}:
    get:
      operationId: getPlanRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A persisted plan run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanRun"
        "404":
          description: The plan run does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The plan run or its parent comparison could not be loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/plan-runs/{run_id}/comparison:
    get:
      operationId: getPlanRunComparison
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A measured comparison with the persisted baseline parent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanComparison"
        "404":
          description: The plan run does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The plan run does not have a baseline parent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: The run or its parent comparison could not be loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/scenarios/{scenario_id}/plan-runs:
    get:
      operationId: listPlanRuns
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: The newest plan runs for the scenario.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanRun"
components:
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message: { type: string }
...
    SiteScenarioIdentity:
      type: object
      additionalProperties: false
      required: [id, name, revision]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        revision: { type: integer, minimum: 1 }
        horizon: { $ref: "#/components/schemas/PlanningHorizon" }
    SiteCapacity:
      type: object
      additionalProperties: false
      required: [solar_kw, wind_kw, battery_kwh, diesel_kw]
      properties:
        solar_kw: { type: number, minimum: 0 }
        wind_kw: { type: number, minimum: 0 }
        battery_kwh: { type: number, minimum: 0 }
        diesel_kw: { type: number, minimum: 0 }
    LatestRunSummary:
      type: object
      additionalProperties: false
      required: [id, scenario_id, planner, status, created_at, summary]
      properties:
        id: { type: string }
        scenario_id: { type: string }
        planner: { type: string, enum: [baseline, wattson] }
        status: { type: string, enum: [computing, complete, infeasible, failed] }
        created_at: { type: string, format: date-time }
        summary: { $ref: "#/components/schemas/PlanSummary" }
    SiteSummary:
      type: object
      additionalProperties: false
      required:
        [
          id,
          name,
          location,
          timezone,
          currency,
          current_scenario,
          capacity,
          service_count,
          commitment_count,
          event_count,
          latest_run,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: string }
        timezone: { type: string }
        currency: { type: string }
        current_scenario: { $ref: "#/components/schemas/SiteScenarioIdentity" }
        capacity: { $ref: "#/components/schemas/SiteCapacity" }
        service_count: { type: integer, minimum: 0 }
        commitment_count: { type: integer, minimum: 0 }
        event_count: { type: integer, minimum: 0 }
        latest_run:
          oneOf:
            - { $ref: "#/components/schemas/LatestRunSummary" }
            - { type: "null" }
    CreateSiteRequest:
      type: object
      additionalProperties: false
      required: [name, location, timezone, currency, solar_capacity_kw, peak_demand_kw]
      properties:
        name: { type: string, minLength: 1 }
        location: { type: string, minLength: 1 }
        timezone: { type: string, minLength: 1 }
        currency: { type: string, pattern: "^[A-Za-z]{3}$" }
        solar_capacity_kw: { type: number, exclusiveMinimum: 0 }
        peak_demand_kw: { type: number, exclusiveMinimum: 0 }
    SiteOverview:
      type: object
      additionalProperties: false
      required:
        [site, current_scenario, commitment_count, event_count, active_event_count, latest_run]
      properties:
        site: { $ref: "#/components/schemas/Site" }
        current_scenario: { $ref: "#/components/schemas/SiteScenarioIdentity" }
        commitment_count: { type: integer, minimum: 0 }
        event_count: { type: integer, minimum: 0 }
        active_event_count: { type: integer, minimum: 0 }
        latest_run:
          oneOf:
            - { $ref: "#/components/schemas/LatestRunSummary" }
            - { type: "null" }
    ForecastLocation:
      type: object
      additionalProperties: false
      required: [name, latitude, longitude, timezone]
      properties:
        name: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        timezone: { type: string }
    ForecastProvider:
      type: object
      additionalProperties: false
      required: [name, url, fetched_at, cache_status]
      properties:
        name: { type: string, const: Open-Meteo }
        url: { type: string, format: uri }
        fetched_at: { type: string, format: date-time }
        cache_status: { type: string, enum: [live, fresh_cache, stale_cache] }
    CurrentWeather:
      type: object
      additionalProperties: false
      required:
        [
          time,
          temperature_c,
          apparent_temperature_c,
          precipitation_mm,
          cloud_cover_percent,
          wind_speed_kph,
          wind_direction_degrees,
          weather_code,
          condition,
          is_day,
        ]
      properties:
        time: { type: string }
        temperature_c: { type: number }
        apparent_temperature_c: { type: number }
        precipitation_mm: { type: number, minimum: 0 }
        cloud_cover_percent: { type: number, minimum: 0, maximum: 100 }
        wind_speed_kph: { type: number, minimum: 0 }
        wind_direction_degrees: { type: number, minimum: 0, maximum: 360 }
        weather_code: { type: integer }
        condition: { type: string }
        is_day: { type: boolean }
...
    WeatherForecast:
      type: object
      additionalProperties: false
      required: [site_id, location, provider, current, hourly, daily, operational_impact]
      properties:
        site_id: { type: string }
        location: { $ref: "#/components/schemas/ForecastLocation" }
        provider: { $ref: "#/components/schemas/ForecastProvider" }
        current: { $ref: "#/components/schemas/CurrentWeather" }
        hourly:
          type: array
          items: { $ref: "#/components/schemas/HourlyWeather" }
        daily:
          type: array
          items: { $ref: "#/components/schemas/DailyWeather" }
        operational_impact: { $ref: "#/components/schemas/OperationalImpact" }
    ScenarioSummary:
      type: object
      additionalProperties: false
      required: [id, name, site_name, location, revision, horizon, contract_count, event_count]
      properties:
        id: { type: string }
        name: { type: string }
        site_name: { type: string }
        location: { type: string }
        revision: { type: integer, minimum: 1 }
        horizon:
          $ref: "#/components/schemas/PlanningHorizon"
        contract_count: { type: integer, minimum: 0 }
        event_count: { type: integer, minimum: 0 }
    Scenario:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - revision
        - id
        - name
        - description
        - site
        - horizon
        - initial_state
        - signals
        - contracts
        - events
        - operating_policy
      properties:
        schema_version: { type: string, const: "1" }
        revision: { type: integer, minimum: 1, readOnly: true }
        id: { type: string }
        name: { type: string }
        description: { type: string }
        site:
          $ref: "#/components/schemas/Site"
        horizon:
          $ref: "#/components/schemas/PlanningHorizon"
        initial_state:
          $ref: "#/components/schemas/InitialState"
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioEvent"
        operating_policy:
          $ref: "#/components/schemas/OperatingPolicy"
    Site:
      type: object
      additionalProperties: false
      required: [id, name, location, timezone, currency, assets, services, connections]
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: string }
        timezone: { type: string }
        currency: { type: string }
        assets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        connections:
          type: array
          description: Directed power connections. The virtual bus node ID is `controller`.
          items:
            $ref: "#/components/schemas/Connection"
...
    ConnectionsRequest:
      type: object
      additionalProperties: false
      required: [connections]
      properties:
        connections:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
...
    PlanningHorizon:
      type: object
      additionalProperties: false
      required: [starts_at, interval_minutes, interval_count]
      properties:
        starts_at: { type: string, format: date-time }
        interval_minutes: { type: integer, const: 15 }
        interval_count: { type: integer, const: 96 }
    InitialState:
      type: object
      additionalProperties: false
      required: [assets]
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AssetState"
...
    Signal:
      oneOf:
        - $ref: "#/components/schemas/RenewableSignal"
        - $ref: "#/components/schemas/ServiceDemandSignal"
        - $ref: "#/components/schemas/FuelDeliverySignal"
      discriminator:
        propertyName: kind
...
    SignalValuesUpdate:
      type: object
      additionalProperties: false
      required: [values]
      properties:
        values:
          $ref: "#/components/schemas/IntervalValues"
    Contract:
      oneOf:
        - $ref: "#/components/schemas/ContinuousPowerContract"
        - $ref: "#/components/schemas/RuntimeByDeadlineContract"
        - $ref: "#/components/schemas/EnergyByDeadlineContract"
      discriminator:
        propertyName: kind
...
    ScenarioEvent:
      oneOf:
        - $ref: "#/components/schemas/RenewableShortfallEvent"
        - $ref: "#/components/schemas/DemandSurgeEvent"
        - $ref: "#/components/schemas/FuelDeliveryDelayEvent"
      discriminator:
        propertyName: type
...
    OperatingPolicy:
      type: object
      additionalProperties: false
      required: [reserve_energy_kwh, assumed_loss_percent]
      properties:
        reserve_energy_kwh: { type: number, minimum: 0 }
        assumed_loss_percent: { type: number, minimum: 0, exclusiveMaximum: 1 }
    PlanningRequest:
      type: object
      additionalProperties: false
      required: [scenario_id, planner, active_event_ids]
      properties:
        scenario_id: { type: string }
        planner: { type: string, enum: [baseline, wattson] }
        active_event_ids:
          type: array
          uniqueItems: true
          items: { type: string }
        parent_run_id: { type: string }
    PlanRun:
      type: object
      additionalProperties: false
      required:
        - id
        - scenario_id
        - scenario_revision
        - scenario_snapshot_hash
        - planner
        - status
        - created_at
        - active_event_ids
        - intervals
        - contract_outcomes
        - decisions
        - summary
      properties:
        id: { type: string }
        scenario_id: { type: string }
        scenario_revision: { type: integer, minimum: 1 }
        scenario_snapshot_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        parent_run_id: { type: string }
        planner: { type: string, enum: [baseline, wattson] }
        status: { type: string, enum: [computing, complete, infeasible, failed] }
        created_at: { type: string, format: date-time }
        active_event_ids:
          type: array
          items: { type: string }
        intervals:
          type: array
          items:
            $ref: "#/components/schemas/PlanInterval"
        contract_outcomes:
          type: array
          items:
            $ref: "#/components/schemas/ContractOutcome"
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        summary:
          $ref: "#/components/schemas/PlanSummary"
        comparison:
          $ref: "#/components/schemas/PlanComparison"
...
    PlanComparison:
      type: object
      additionalProperties: false
      required:
        - baseline_run_id
        - candidate_run_id
        - summary
        - contracts
        - services
        - renewables
        - batteries
        - generators
        - intervals
      properties:
        baseline_run_id: { type: string }
        candidate_run_id: { type: string }
        summary: { $ref: "#/components/schemas/ComparisonSummary" }
        contracts:
          type: array
          items: { $ref: "#/components/schemas/ContractComparison" }
        services:
          type: array
          items: { $ref: "#/components/schemas/ServiceComparison" }
        renewables:
          type: array
          items: { $ref: "#/components/schemas/RenewableComparison" }
        batteries:
          type: array
          items: { $ref: "#/components/schemas/BatteryComparison" }
        generators:
          type: array
          items: { $ref: "#/components/schemas/GeneratorComparison" }
        intervals:
          type: array
          items: { $ref: "#/components/schemas/IntervalComparison" }
...